# contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/content/upload": {
      "post": {
        "operationId": "ContentController_upload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadContentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload image/video/PDF content.",
        "tags": [
          "content"
        ]
      }
    },
    "/content/url": {
      "post": {
        "operationId": "ContentController_createUrl",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUrlContentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create URL content.",
        "tags": [
          "content"
        ]
      }
    },
    "/content/text": {
      "post": {
        "operationId": "ContentController_createText",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTextContentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a text announcement.",
        "tags": [
          "content"
        ]
      }
    },
    "/content": {
      "get": {
        "operationId": "ContentController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "meta"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ContentDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PageMetaDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List content (filter by type/status/orientation/tag/expiry; sortable).",
        "tags": [
          "content"
        ]
      }
    },
    "/content/usage": {
      "get": {
        "operationId": "ContentController_usage",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Storage usage + content counts for the library.",
        "tags": [
          "content"
        ]
      }
    },
    "/content/bulk/archive": {
      "post": {
        "operationId": "ContentController_bulkArchive",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkContentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "content"
        ]
      }
    },
    "/content/bulk/unarchive": {
      "post": {
        "operationId": "ContentController_bulkUnarchive",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkContentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "content"
        ]
      }
    },
    "/content/bulk/trash": {
      "post": {
        "operationId": "ContentController_bulkTrash",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkContentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "content"
        ]
      }
    },
    "/content/bulk/restore": {
      "post": {
        "operationId": "ContentController_bulkRestore",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkContentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "content"
        ]
      }
    },
    "/content/bulk/tags": {
      "post": {
        "operationId": "ContentController_bulkTags",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkContentTagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "content"
        ]
      }
    },
    "/content/trash/purge": {
      "post": {
        "operationId": "ContentController_purge",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurgedResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Permanently remove this company’s trash older than 14 days.",
        "tags": [
          "content"
        ]
      }
    },
    "/content/{id}": {
      "get": {
        "operationId": "ContentController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Content detail.",
        "tags": [
          "content"
        ]
      },
      "patch": {
        "operationId": "ContentController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update content metadata (title/description/orientation/expiry/tags/...).",
        "tags": [
          "content"
        ]
      }
    },
    "/content/{id}/preview": {
      "get": {
        "operationId": "ContentController_preview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "A signed preview URL (files) or the URL/text payload.",
        "tags": [
          "content"
        ]
      }
    },
    "/content/{id}/replace": {
      "post": {
        "operationId": "ContentController_replace",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Replace the uploaded file (same type).",
        "tags": [
          "content"
        ]
      }
    },
    "/content/{id}/archive": {
      "post": {
        "operationId": "ContentController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "content"
        ]
      }
    },
    "/content/{id}/unarchive": {
      "post": {
        "operationId": "ContentController_unarchive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "content"
        ]
      }
    },
    "/content/{id}/trash": {
      "post": {
        "operationId": "ContentController_trash",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "content"
        ]
      }
    },
    "/content/{id}/restore": {
      "post": {
        "operationId": "ContentController_restore",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "content"
        ]
      }
    },
...
  },
  "info": {
    "title": "Wizer Signage API",
    "description": "Wizer Signage — multi-tenant digital signage SaaS platform REST API.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PageMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 25
          },
          "total": {
            "type": "number",
            "example": 137,
            "description": "Total matching rows, ignoring pagination."
          },
          "totalPages": {
            "type": "number",
            "example": 6
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ]
      },
...
      "TagDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "companyId": {
            "type": "string",
            "description": "Owning tenant. Derived from the caller's token, never the body."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "color": {
            "type": "object",
            "nullable": true,
            "example": "#2F80ED"
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "type": {
            "type": "string",
            "enum": [
              "CONTENT",
              "SCREEN",
              "BOTH"
            ],
            "description": "Which resources the tag may be applied to."
          }
        },
        "required": [
          "id",
          "companyId",
          "createdAt",
          "updatedAt",
          "name"
        ]
      },
...
      "UploadContentDto": {
        "type": "object",
        "properties": {}
      },
      "ContentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "companyId": {
            "type": "string",
            "description": "Owning tenant. Derived from the caller's token, never the body."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "object",
            "nullable": true
          },
          "type": {
            "type": "string",
            "enum": [
              "IMAGE",
              "VIDEO",
              "PDF",
              "URL",
              "TEXT"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "ARCHIVED",
              "TRASHED"
            ]
          },
          "orientation": {
            "type": "string",
            "enum": [
              "LANDSCAPE",
              "PORTRAIT",
              "UNKNOWN"
            ]
          },
          "url": {
            "type": "object",
            "nullable": true,
            "description": "External address for URL content."
          },
          "textBody": {
            "type": "object",
            "nullable": true,
            "description": "Body for TEXT content."
          },
          "textStyle": {
            "type": "object",
            "nullable": true
          },
          "fileSize": {
            "type": "string",
            "nullable": true,
            "description": "Bytes, as a STRING — a 64-bit column serialised via BigInt.toJSON."
          },
          "mimeType": {
            "type": "object",
            "nullable": true
          },
          "originalFileName": {
            "type": "object",
            "nullable": true
          },
          "pageCount": {
            "type": "object",
            "nullable": true,
            "description": "Pages in a PDF."
          },
          "durationSeconds": {
            "type": "object",
            "nullable": true,
            "description": "Default dwell time when scheduled."
          },
          "expiresAt": {
            "type": "object",
            "nullable": true,
            "format": "date-time"
          },
          "isExpired": {
            "type": "boolean",
            "description": "Derived at read time from expiresAt — there is no such column."
          },
          "archivedAt": {
            "type": "object",
            "nullable": true,
            "format": "date-time"
          },
          "trashedAt": {
            "type": "object",
            "nullable": true,
            "format": "date-time"
          },
          "createdById": {
            "type": "object",
            "nullable": true
          },
          "updatedById": {
            "type": "object",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagDto"
            }
          }
        },
        "required": [
          "id",
          "companyId",
          "createdAt",
          "updatedAt",
          "title",
          "type",
          "status",
          "orientation",
          "isExpired",
          "tags"
        ]
      },
      "CreateUrlContentDto": {
        "type": "object",
        "properties": {}
      },
      "CreateTextContentDto": {
        "type": "object",
        "properties": {}
      },
      "BulkContentDto": {
        "type": "object",
        "properties": {}
      },
      "BulkContentTagDto": {
        "type": "object",
        "properties": {}
      },
      "PurgedResponseDto": {
        "type": "object",
        "properties": {
          "purged": {
            "type": "number",
            "example": 3,
            "description": "How many trashed items were permanently removed."
          }
        },
        "required": [
          "purged"
        ]
      },
      "UpdateContentDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}